+++ src/Codecat/Plugins/PluginRegistry.cs
@@ -241,6 +241,22 @@ internal static class PluginRegistry
                 ignoredDirectories: ignored),
 
             new ExtensionPlugin(
+                "shader",
+                new Dictionary<string, string>
+                {
+                    [".wgsl"] = "wgsl"
+                },
+                ignoredDirectories: ignored),
+
+            new ExtensionPlugin(
+                "script",
+                new Dictionary<string, string>
+                {
+                    [".rhai"] = "rhai"
+                },
+                ignoredDirectories: ignored),
+
+            new ExtensionPlugin(
                 "php",
                 new Dictionary<string, string> { [".php"] = "php", [".phtml"] = "php" },
                 new Dictionary<string, string>
@@ -428,7 +444,6 @@ internal static class PluginRegistry
                     [".json"] = "json",
                     [".yaml"] = "yaml",
                     [".yml"] = "yaml",
-                    [".toml"] = "toml",
                     [".ini"] = "ini",
                     [".conf"] = "conf",
                     [".config"] = "xml",
@@ -451,6 +466,14 @@ internal static class PluginRegistry
                     ["codegraph.toml"] = "toml"
                 },
                 ignored),
+
+            new ExtensionPlugin(
+                "toml",
+                new Dictionary<string, string>
+                {
+                    [".toml"] = "toml"
+                },
+                ignoredDirectories: ignored),
 
             new ExtensionPlugin(
                 "docs",